Release v2.0.0

+++ data/src/main/java/daily/dayo/data/datasource/remote/bookmark/BookmarkRequest.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class CreateBookmarkRequest(
     @SerializedName("postId")
-    val postId: Int
+    val postId: Long
 )
