+++ domain/src/main/java/daily/dayo/domain/usecase/comment/RequestDeletePostCommentUseCase.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class RequestDeletePostCommentUseCase @Inject constructor(
     private val commentRepository: CommentRepository
 ) {
-    suspend operator fun invoke(commentId: Int) =
+    suspend operator fun invoke(commentId: Long) =
         commentRepository.requestDeletePostComment(commentId)
 }
